refactor: checkAccountOwnership and validateSignature to middleware/near.js, added routes to initialize and validate code for temp accounts (used in wallet to verify recovery method prior to creating account on chain)
